refactor: analytics feature
